Comments out making new amphitheatre object a child and moves scene
Objects spawned by amphitheatre blocks are better not being children when the block has a limited lifespan because this attaches the lifespan of the block to the child. Also, this makes it so that the block spins out of the scene 3 seconds after spawning, which is technically a bug and while it could be fixed, it looks nice and is now a feature.